fix norm
